+++ src/Mailtrap.Abstractions/GlobalUsings.cs
@@ -68,6 +68,8 @@ global using Mailtrap.Projects.Responses;
 global using Mailtrap.SendingDomains;
 global using Mailtrap.SendingDomains.Models;
 global using Mailtrap.SendingDomains.Requests;
+global using Mailtrap.Stats;
+global using Mailtrap.Stats.Models;
 global using Mailtrap.Suppressions;
 global using Mailtrap.Suppressions.Models;
 global using Mailtrap.TestingMessages;