+++ tests/Mailtrap.UnitTests/Stats/StatsResourceTests.cs
@@ -1,0 +1,52 @@
+namespace Mailtrap.UnitTests.Stats;
+
+
+[TestFixture]
+internal sealed class StatsResourceTests
+{
+    private readonly IRestResourceCommandFactory _commandFactoryMock = Mock.Of<IRestResourceCommandFactory>();
+    private readonly Uri _resourceUri = EndpointsTestConstants.ApiDefaultUrl
+        .Append(
+            UrlSegmentsTestConstants.ApiRootSegment,
+            UrlSegmentsTestConstants.AccountsSegment)
+        .Append(TestContext.CurrentContext.Random.NextLong())
+        .Append(UrlSegmentsTestConstants.StatsSegment);
+
+
+    #region Constructor
+
+    [Test]
+    public void Constructor_ShouldThrowArgumentNullException_WhenCommandFactoryIsNull()
+    {
+        // Act
+        var act = () => new StatsResource(null!, _resourceUri);
+
+        // Assert
+        act.Should().Throw<ArgumentNullException>();
+    }
+
+    [Test]
+    public void Constructor_ShouldThrowArgumentNullException_WhenUriIsNull()
+    {
+        // Act
+        var act = () => new StatsResource(_commandFactoryMock, null!);
+
+        // Assert
+        act.Should().Throw<ArgumentNullException>();
+    }
+
+    [Test]
+    public void ResourceUri_ShouldBeInitializedProperly()
+    {
+        // Arrange
+        var client = CreateResource();
+
+        // Assert
+        client.ResourceUri.Should().Be(_resourceUri);
+    }
+
+    #endregion
+
+
+    private StatsResource CreateResource() => new(_commandFactoryMock, _resourceUri);
+}